Updated URL de-telemetry to work with non-spooked YouTube URLs
Initially I assumed that YouTube _only_ shares `youtu.be` URLs, but it turns out they also share `youtube.com/watch?v=` (normal) URLs (albeit with an `si` query string for some kind of spook.
Updated the function to use a regex to evaluate whether or not the input URL is a `youtube.com/watch` (good) URL, and in that case just remove the query strings that aren't the `v=<video id>` bit. Otherwise, it will just do the de-spooking as it used to.